requirements.txt
websockets>=12.0
numpy>=1.24.0
Pillow>=10.0.0
pygame>=2.5.0
